Checking NPE in Service

+++ src/main/java/ru/webmarket/service/impl/CategoryServiceImpl.java
@@ -20,34 +20,42 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public void addCategory(CategoryDTO categoryDTO) {
         Category category = CategoryConvector.dtoToEntity(categoryDTO);
-        categoryRepository.save(category);
+        if (category != null) categoryRepository.save(category);
     }
 
     @Override
     public CategoryDTO getCategory(Long id) {
         CategoryDTO categoryDTO = CategoryConvector.entityToDto(categoryRepository.findOne(id));
+        if (categoryDTO == null) return null;
         return categoryDTO;
     }
 
     @Override
     public void editCategory(CategoryDTO toEditCategoryDTO) {
         Category category = categoryRepository.findOne(toEditCategoryDTO.getId());
-
-        category.setName(toEditCategoryDTO.getName());
+        if (category != null) {
+            category.setName(toEditCategoryDTO.getName());
+        }
     }
 
     @Override
     public void deleteCategory(Long id) {
-        categoryRepository.delete(id);
+        Category category = categoryRepository.findOne(id);
+        if (category != null) categoryRepository.delete(id);
     }
 
     @Override
     public List<CategoryDTO> getAllCategories() {
-        return CategoryConvector.entityToDto(categoryRepository.findAll());
+        List<CategoryDTO> categoryDTOS = new ArrayList<>();
+        categoryDTOS = CategoryConvector.entityToDto(categoryRepository.findAll());
+        if (categoryDTOS.isEmpty()) return null;
+        return categoryDTOS;
     }
 
     @Override
     public CategoryDTO findByName(String name) {
-        return CategoryConvector.entityToDto(categoryRepository.findByName(name));
+        Category category = categoryRepository.findByName(name);
+        if (category == null) return null;
+        return CategoryConvector.entityToDto(category);
     }
 }
