+++ src/main/java/ru/webmarket/service/impl/ProductServiceImpl.java
@@ -10,6 +10,7 @@ import ru.webmarket.entity.dto.ProductDTO;
 import ru.webmarket.repository.ProductRepository;
 import ru.webmarket.service.ProductService;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -21,40 +22,53 @@ public class ProductServiceImpl implements ProductService{
 
     @Override
     public void addProduct(ProductDTO productDTO) {
-        productRepository.save(ProductConverter.dtoToEntity(productDTO));
+        Product product = productRepository.findOne(productDTO.getId());
+        if (product != null) productRepository.save(ProductConverter.dtoToEntity(productDTO));
     }
 
     @Override
     public ProductDTO getProduct(Long id) {
-        return ProductConverter.entityToDto(productRepository.findOne(id));
+        Product product = productRepository.findOne(id);
+        if (product == null) return null;
+        return ProductConverter.entityToDto(product);
     }
 
     @Override
     public void editProduct(ProductDTO productDTO) {
         Product product = productRepository.findOne(productDTO.getId());
-
-        product.setName(productDTO.getName());
-        product.setCategories(CategoryConvector.dtoToEntity(productDTO.getCategories()));
-        product.setPrice(productDTO.getPrice());
+        if (product != null) {
+            product.setName(productDTO.getName());
+            product.setCategories(CategoryConvector.dtoToEntity(productDTO.getCategories()));
+            product.setPrice(productDTO.getPrice());
+        }
     }
 
     @Override
     public void deleteProduct(Long id) {
-        productRepository.delete(id);
+        Product product = productRepository.findOne(id);
+        if (product != null) productRepository.delete(id);
     }
 
     @Override
     public List<ProductDTO> getAllProduct() {
-        return ProductConverter.entityToDto(productRepository.findAll());
+        List<ProductDTO> productDTOS = new ArrayList<>();
+        productDTOS = ProductConverter.entityToDto(productRepository.findAll());
+        if (productDTOS == null) return null;
+        return productDTOS;
     }
 
     @Override
     public List<CategoryDTO> getAllCategory(ProductDTO productDTO) {
-        return productDTO.getCategories();
+        List<CategoryDTO> categoryDTOS = new ArrayList<>();
+        categoryDTOS = productDTO.getCategories();
+        if (categoryDTOS.isEmpty()) return null;
+        return categoryDTOS;
     }
 
     @Override
     public ProductDTO findByName(String name) {
-        return ProductConverter.entityToDto(productRepository.findByName(name));
+        Product product = productRepository.findByName(name);
+        if (product == null) return null;
+        return ProductConverter.entityToDto(product);
     }
 }
